+++ src/Objects/Collection.php
@@ -18,7 +18,7 @@ class Collection extends AbstractList
      * @param string $key
      * @param string $ref
      * 
-     * @return KeyValue Construct a collection item instance. A KeyValue or a custom class you set with setChildClass().
+     * @return KeyValueInterface Construct a collection item instance. A KeyValue or a custom class you set with setChildClass().
      */
     public function item($key = null, $ref = null)
     {