+++ src/Objects/ReusableObjectInterface.php
@@ -1,6 +1,8 @@
 <?php
 namespace andrefelipe\Orchestrate\Objects;
 
+use andrefelipe\Orchestrate\ClientInterface;
+
 /**
  * Defines an object as being reusable. We can reset and init the object without creating a new instance.
  */
@@ -18,4 +20,20 @@ interface ReusableObjectInterface
      * @param array $data
      */
     public function init(array $data);
+
+    /**
+     * Get current client instance, either of Application or Client class.
+     * 
+     * @param boolean $required
+     * 
+     * @return ClientInterface
+     */
+    public function getClient($required = false);
+
+    /**
+     * Set the client which the object will use to make API requests.
+     * 
+     * @param ClientInterface $client
+     */
+    public function setClient(ClientInterface $client);
 }